migrate NullCoalesce Rule to IssetRule
